Saved meals are now filtered from returned saved recipes

+++ bandits-api/src/main/java/com/bandits/api/bandits_api/repositories/ReviewPlanRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
 public interface ReviewPlanRepository extends JpaRepository<Meal, Integer> {
-    List<Meal> findByRecipe_User_Id(Long userId);
+    List<Meal> findByRecipe_User_Id(Integer userId);
 }
